steaming_ai/requirements.txt
requests
groq
elevenlabs
Flask
Flask-SocketIO
python-dotenv
simple-websocket
gtts 